Account for not providing environment on non-JSC machine
This accounts for the complete truth table of (is var set?) and (are we
on JSC?).  If neither is the case, the scripts aborts.

Furthermore, the abort message format is made consistent with another
abort location (where I did not add a suggestion as providing model
component parameters is already in the README).

Finally, I made messages for "things about to be done" contrast to
aborts by making them present continuous tense and adding an ellipsis.