Use a better heuristic for the hostname to SYSTEMNAME function
The longest FQDN typically found for clusters is
NODE.SYSTEMNAME.ORGANISATION.COUNTRY.  Usually only part of this.
The node name always ends with a digit (as part of a sequence number of
the login node).  The first part without a digit is selected.

This breaks sometimes a bit for numbered clients.  E.g. gfi201.uib.no
would yield 'uib' (instead of the previous 'gfi' or the preferred
'gfi201').

Simply selecting part two of the long domain name, but that breaks then
also for clients (laptops and workstations) that have an alphabetic
short hostname.

Also simplified removal of prefixed and trailing spaces.