Normalised the use of SYSTEMNAME
The variable $MACHINE is not needed and hostname(1) not desired.  I left
the hostname command on line 30, because SYSTEMNAME can be empty (this
is unlikely and should not break anything).

Where the old $MACHINE==`hostname` needed to be checked against the
nondescript "login", with the updated $SYSTEMNAME it must now be checked
against the more pleasant "marvin" to make Marvin less depressed.