Account for case of only parts ending with a digit
Then it defaults to the first part (equivallent to hostname -s).

To explain further commit c796dae, in that host naming scheme (which
should be the standard, hence part of the heuristic) the second part is
the cluster name.  However, the code allows complexities like
node01.rack1.mycluster[.orgname.org].